+++ src/file_system/directory.rs
@@ -3,9 +3,9 @@ pub const TYPE_FILE: u16 = 2;
 pub const TYPE_DEVICE: u16 = 3;
 
 pub struct FileStatus {
-    pub dev: usize,
-    pub ino: usize,
+    pub dev: u32,
+    pub ino: u32,
     pub types: u16,
     pub nlink: u16,
-    pub size: usize,
+    pub size: u32,
 }
